1.更新echart api：获取商品数量 2.小程序以前需要userId的现在变为从token中获取

+++ dgutshop-admin-api/src/main/java/org/dgutstu/dgutshop/admin/controller/AdminAdminController.java
@@ -106,7 +106,7 @@ public class AdminAdminController {
 
     @PreAuthorize("hasRole('ADMIN')")
     @PostMapping("/delete")
-    public Object delete(@RequestParam Long id) {
+    public Object delete(@RequestParam Integer id) {
         if (id == null) {
             return ResponseUtil.badArgument();
         }
@@ -142,7 +142,7 @@ public class AdminAdminController {
             return error;
         }
 
-        Long adminId = admin.getId();
+        Integer adminId = admin.getId();
         if (adminId == null) {
             return ResponseUtil.badArgument();
         }
