+++ dgutshop-db/src/main/java/org/dgutstu/dgutshop/db/service/DgutshopRoleService.java
@@ -79,11 +79,11 @@ public class DgutshopRoleService {
      * 删除第id个角色
      * @param id
      */
-    public void deleteById(Long id){
+    public void deleteById(Integer id){
         roleMapper.logicalDeleteByPrimaryKey(id);
     }
 
-    public DgutshopRole findById(Long id){
+    public DgutshopRole findById(Integer id){
         return roleMapper.selectByPrimaryKey(id);
     }
 }
